fix: don't load self convo's messages into memory
